moving all the robotlomoction translators to drake/lcm/robotlocomotion_translators. adding cmake stuff.
