+++ drake/examples/QPInverseDynamicsForHumanoids/system/BUILD
@@ -84,6 +84,7 @@ drake_cc_library(
         ":joint_level_controller_system",
         ":manipulator_plan_eval_system",
         ":qp_controller_system",
+        "//drake/examples/QPInverseDynamicsForHumanoids:humanoid_status_translator",
         "//drake/multibody/parsers",
         "//drake/systems/controllers:state_feedback_controller_interface",
         "//drake/systems/framework",
